차트페이지 외에 차트가 보이지 않도록  수정
