+++ client/src/components/chart/OptionList.tsx
@@ -6,35 +6,6 @@ import React, { useState, useEffect } from 'react';
 import { Radio } from 'antd';
 
 
-const sample = ["x", "2021-11-10", "2021-11-12", "2021-11-12", "2021-11-14", "2021-11-22"]
-const dibs = {"total":10, "2021-11-10":1, "2021-11-12":2}
-
-function isNumeric(data : string) : boolean {
-  return !isNaN(Number(data));
-}
-
-function dibsDatetoMonth(data: Object) {
-  var month: any[] = [];
-  console.log(Object.keys(data));
-  for (const key of Object.keys(data) as (keyof typeof data)[]) {
-    console.log(data[key]);
-    if (isNumeric(key[0])) {
-      if (!month.includes(key)) {
-        console.log(key);
-        // 에러 발생
-        const tempYear = ((key as unknown) as Date).getFullYear();
-        const tempMonth = (key as unknown as Date).getMonth();
-        const tempDate = tempYear + "-" + tempMonth;
-        console.log(tempDate)
-      }
-    }
-  }
-}
-
-function dibsDatetoDay(data: Object) {
-  
-}
-
 interface OptionProps {
   optionID : string;
 }
@@ -44,15 +15,20 @@ const OptionList = (props: OptionProps) => {
   // const [OptionList, setOptionList] = useState([]);
   // const [productName] = useRecoilState(productState);
 
-  const options = props.optionID
+  const options = props.optionID;
   const [option, setOption] = useState(1);
 
 
   const handleInputChange = (value: number) => {
-    if (value == 2) 
-      dibsDatetoDay(dibs);
-    if (value == 3) 
-      dibsDatetoMonth(dibs);
+    if (value == 1) {
+      console.log("1")
+    }
+    if (value == 2) {
+      console.log("2");
+    }
+    if (value == 3) {
+      console.log("3");
+    }
     setOption(value);
   };
